🤖 tests: fold assisted-review aggregator tests into main suite
Move the four review_pane_update tests into StreamingMessageAggregator.test.ts
to keep the alphabetical file ordering stable. Bun runs test files in name
order, and inserting a new top-level test file shifted that order enough to
expose a few latent test-pollution failures in unrelated files (GitStatusStore
losing window listeners, etc.). Inlining the tests preserves the original
order and significantly drops the noisy-flake count on the full suite.
